6-size.c

+++ 0x00-hello_world/6-size.c
@@ -9,10 +9,10 @@ int b;
 long int c;
 long long int d;
 float e;
-printf("Size of a char: %u byte(s)\n", sizeof(char));
-printf("Sizeof of an int: %u byte(s)\n", sizeof(b));
-printf("Sizeof of a long int: %u byte(s)\n", sizeof(c));
-printf("Size of a long long int: %u byte(s)\n", sizeof(d));
-printf("Size of a float: %u byte(s)\n", sizeof(e));
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
+printf("Sizeof of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
+printf("Sizeof of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
+printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(e));
 return (0);
 }
